requirements.txt
selenium>=4.21.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.12.3
lxml>=5.2.2
fastmcp>=0.9.0
requests>=2.31.0
python-pptx>=0.6.21
